versão com estatisticas no lugar da escrita em arquivo implementada

+++ src/main/java/jv/records/Document.java
@@ -1,4 +1,4 @@
 package jv.records;
 
 import java.util.Map;
-public record Document(int id, Map<String, Long> counts, long n_terms) {}
+public record Document(long id, Map<String, Long> counts, long n_terms) {}
